Move ResultNameResolver
